feat: extract IReturnTypeGenerator, IMethodAttributeGenerator, IMethodSignatureGenerator, and IMethodGenerator
InterfaceGenerator reduced from 450 to 253 lines by extracting:
- ReturnTypeGenerator: return type derivation, IApiResponse wrapping, Task/IObservable
- MethodAttributeGenerator: Accept/Content-Type/Authorization headers, [Multipart], [Obsolete]
- MethodSignatureGenerator: parameter aggregation (delegates to ParameterAggregator)
- MethodGenerator: facade composing the three sub-generators

Public API unchanged. All existing tests continue to work.
