+++ src/Refitter.Core/InterfaceGenerator.cs
@@ -1,5 +1,4 @@
 using System.Text;
-using System.Text.RegularExpressions;
 using NSwag;
 using NSwag.CodeGeneration.CSharp.Models;
 
@@ -8,14 +7,12 @@ namespace Refitter.Core;
 internal class InterfaceGenerator
 {
     private const string Separator = "    ";
-    private static readonly Regex HttpResponseMessageTypeRegex = new("(Task|IObservable)<HttpResponseMessage>", RegexOptions.Compiled, TimeSpan.FromSeconds(1));
-    private static readonly Regex ApiResponseTypeRegex = new("(Task|IObservable)<(I)?ApiResponse(<[\\w<>]+>)?>", RegexOptions.Compiled, TimeSpan.FromSeconds(1));
 
     private readonly RefitGeneratorSettings settings;
     private readonly OpenApiDocument document;
     private readonly CustomCSharpClientGenerator generator;
     private readonly XmlDocumentationGenerator docGenerator;
-    private readonly IParameterExtractor parameterExtractor;
+    private readonly IMethodGenerator methodGenerator;
 
     internal InterfaceGenerator(
         RefitGeneratorSettings settings,
@@ -23,12 +20,26 @@ internal class InterfaceGenerator
         CustomCSharpClientGenerator generator,
         XmlDocumentationGenerator docGenerator,
         IParameterExtractor? parameterExtractor = null)
+        : this(settings, document, generator, docGenerator,
+            new MethodGenerator(
+                new ReturnTypeGenerator(settings, generator),
+                new MethodAttributeGenerator(settings, document),
+                new MethodSignatureGenerator(settings, parameterExtractor ?? new ParameterAggregator())))
+    {
+    }
+
+    internal InterfaceGenerator(
+        RefitGeneratorSettings settings,
+        OpenApiDocument document,
+        CustomCSharpClientGenerator generator,
+        XmlDocumentationGenerator docGenerator,
+        IMethodGenerator methodGenerator)
     {
         this.settings = settings;
         this.document = document;
         this.generator = generator;
         this.docGenerator = docGenerator;
-        this.parameterExtractor = parameterExtractor ?? new ParameterAggregator();
+        this.methodGenerator = methodGenerator;
         generator.BaseSettings.OperationNameGenerator = new OperationNameGenerator(document, settings);
     }
 
@@ -158,7 +169,7 @@ internal class InterfaceGenerator
             return (string.Empty, string.Empty);
         }
 
-        var returnType = GetTypeName(operation);
+        var returnType = methodGenerator.GenerateReturnType(operation);
         var verb = op.Verb.CapitalizeFirstCharacter();
         var baseOperationName = GetBaseOperationName(op);
         var rawMethodName = partitioning.GetMethodName(op, interfaceName, baseOperationName);
@@ -167,22 +178,24 @@ internal class InterfaceGenerator
 
         var dynamicQuerystringParameterType = partitioning.GetDynamicQuerystringParameterType(interfaceName, methodName);
         var operationModel = generator.CreateOperationModel(operation);
-        var parameters = parameterExtractor.ExtractParameters(operationModel, operation, settings, dynamicQuerystringParameterType, out var operationDynamicQuerystringParameters).ToList();
+
+        var (parametersString, parameters, operationDynamicQuerystringParameters) =
+            methodGenerator.GenerateMethodSignature(operationModel, operation, dynamicQuerystringParameterType);
 
         var hasDynamicQuerystringParameter = !string.IsNullOrWhiteSpace(operationDynamicQuerystringParameters);
-        var parametersString = string.Join(", ", parameters);
         var hasApizrRequestOptionsParameter = settings.ApizrSettings?.WithRequestOptions == true;
         var hasCancellationToken = settings.UseCancellationTokens && !hasApizrRequestOptionsParameter;
-        var isApiResponseType = IsApiResponseType(returnType);
+        var isApiResponseType = methodGenerator.IsApiResponseType(returnType);
 
         if (settings.GenerateXmlDocCodeComments)
         {
             docGenerator.AppendMethodDocumentation(operationModel, isApiResponseType, hasDynamicQuerystringParameter, hasApizrRequestOptionsParameter, hasCancellationToken, code);
         }
 
-        GenerateObsoleteAttribute(operation, code);
-        GenerateForMultipartFormData(operationModel, code);
-        GenerateHeaders(operation, operationModel, code);
+        foreach (var attribute in methodGenerator.GenerateMethodAttributes(operation, operationModel))
+        {
+            code.AppendLine($"{Separator}{Separator}{attribute}");
+        }
 
         code.AppendLine($"{Separator}{Separator}[{verb}(\"{op.Path}\")]")
             .AppendLine($"{Separator}{Separator}{returnType} {methodName}({parametersString});")
@@ -195,14 +208,17 @@ internal class InterfaceGenerator
                 docGenerator.AppendMethodDocumentation(operationModel, isApiResponseType, false, hasApizrRequestOptionsParameter, hasCancellationToken, code);
             }
 
-            GenerateObsoleteAttribute(operation, code);
-            GenerateForMultipartFormData(operationModel, code);
-            GenerateHeaders(operation, operationModel, code);
+            foreach (var attribute in methodGenerator.GenerateMethodAttributes(operation, operationModel))
+            {
+                code.AppendLine($"{Separator}{Separator}{attribute}");
+            }
 
-            parametersString = string.Join(", ", parameters.Where(parameter => !parameter.Contains("?")));
+            var nonOptionalParametersString = string.Join(
+                ", ",
+                parameters.Where(parameter => !parameter.Contains("?")));
 
             code.AppendLine($"{Separator}{Separator}[{verb}(\"{op.Path}\")]")
-                .AppendLine($"{Separator}{Separator}{returnType} {methodName}({parametersString});")
+                .AppendLine($"{Separator}{Separator}{returnType} {methodName}({nonOptionalParametersString});")
                 .AppendLine();
         }
 
@@ -215,219 +231,6 @@ internal class InterfaceGenerator
             .BaseSettings
             .OperationNameGenerator
             .GetOperationName(document, op.Path, op.Verb, op.Operation);
-    }
-
-    private static string TrimImportedNamespaces(string returnTypeParameter) =>
-        returnTypeParameter.StartsWith("System.Collections.Generic.", StringComparison.OrdinalIgnoreCase)
-            ? returnTypeParameter.Replace("System.Collections.Generic.", string.Empty)
-            : returnTypeParameter;
-
-    private string GetTypeName(OpenApiOperation operation)
-    {
-        if (settings.ResponseTypeOverride.TryGetValue(operation.OperationId, out var type))
-        {
-            return type is null or "void"
-                ? GetAsyncOperationType(true)
-                : $"{GetAsyncOperationType(false)}<{TrimImportedNamespaces(type)}>";
-        }
-
-        if (IsFileStreamResponse(operation))
-        {
-            return $"{GetAsyncOperationType(false)}<HttpResponseMessage>";
-        }
-
-        var successCodes = new[] { "200", "201", "203", "206" };
-        var returnTypeParameter = successCodes
-            .Where(operation.Responses.ContainsKey)
-            .Select(code => GetTypeName(code, operation))
-            .FirstOrDefault();
-
-        if (returnTypeParameter == null && operation.Responses.ContainsKey("2XX"))
-        {
-            returnTypeParameter = GetTypeName("2XX", operation);
-        }
-
-        if (returnTypeParameter == null && operation.Responses.ContainsKey("default"))
-        {
-            returnTypeParameter = GetTypeName("default", operation);
-        }
-
-        return GetReturnType(returnTypeParameter);
-    }
-
-    private static bool IsFileStreamResponse(OpenApiOperation operation)
-    {
-        var successCodes = new[] { "200", "201", "203", "206", "2XX" };
-
-        foreach (var code in successCodes)
-        {
-            if (!operation.Responses.TryGetValue(code, out var apiResponse))
-                continue;
-
-            var response = apiResponse.ActualResponse;
-
-            if (response.Content?.Any() != true)
-                continue;
-
-            foreach (var contentEntry in response.Content)
-            {
-                if (IsFileContentType(contentEntry.Key))
-                {
-                    var schema = contentEntry.Value?.Schema;
-                    if (schema?.Format == "binary" || schema?.Type == NJsonSchema.JsonObjectType.File)
-                        return true;
-                }
-            }
-        }
-
-        return false;
-    }
-
-    private static bool IsFileContentType(string contentType)
-    {
-        return
-            contentType.StartsWith("application/octet-stream", StringComparison.OrdinalIgnoreCase) ||
-            contentType.StartsWith("application/pdf", StringComparison.OrdinalIgnoreCase) ||
-            contentType.StartsWith("application/vnd", StringComparison.OrdinalIgnoreCase) ||
-            contentType.StartsWith("image/", StringComparison.OrdinalIgnoreCase) ||
-            contentType.StartsWith("video/", StringComparison.OrdinalIgnoreCase) ||
-            contentType.StartsWith("audio/", StringComparison.OrdinalIgnoreCase) ||
-            contentType.StartsWith("application/zip", StringComparison.OrdinalIgnoreCase) ||
-            contentType.StartsWith("application/gzip", StringComparison.OrdinalIgnoreCase) ||
-            (contentType.StartsWith("application/x-", StringComparison.OrdinalIgnoreCase) &&
-             !contentType.StartsWith("application/x-www-form-urlencoded", StringComparison.OrdinalIgnoreCase));
-    }
-
-    private string GetTypeName(string code, OpenApiOperation operation)
-    {
-        var schema = operation.Responses[code].ActualResponse.Schema;
-        var typeName = generator.GetTypeName(schema, false, null);
-
-        if (!string.IsNullOrWhiteSpace(settings.CodeGeneratorSettings?.ArrayType) &&
-            schema?.Type == NJsonSchema.JsonObjectType.Array)
-        {
-            typeName = typeName
-                .Replace("ICollection", settings.CodeGeneratorSettings!.ArrayType)
-                .Replace("IEnumerable", settings.CodeGeneratorSettings!.ArrayType);
-        }
-
-        return typeName;
-    }
-
-    private string GetReturnType(string? returnTypeParameter)
-    {
-        return returnTypeParameter is null or "void"
-            ? GetDefaultReturnType()
-            : GetConfiguredReturnType(returnTypeParameter);
-    }
-
-    private string GetDefaultReturnType()
-    {
-        var asyncType = GetAsyncOperationType(true);
-        return settings.ReturnIApiResponse
-            ? $"{asyncType}<IApiResponse>"
-            : asyncType;
-    }
-
-    private string GetConfiguredReturnType(string returnTypeParameter)
-    {
-        var asyncType = GetAsyncOperationType(false);
-        return settings.ReturnIApiResponse
-            ? $"{asyncType}<IApiResponse<{TrimImportedNamespaces(returnTypeParameter)}>>"
-            : $"{asyncType}<{TrimImportedNamespaces(returnTypeParameter)}>";
-    }
-
-    private string GetAsyncOperationType(bool withVoidReturnType)
-    {
-        var type = withVoidReturnType ? "<Unit>" : string.Empty;
-        return settings.ReturnIObservable
-            ? "IObservable" + type
-            : "Task";
-    }
-
-    private static bool IsApiResponseType(string typeName)
-    {
-        return HttpResponseMessageTypeRegex.IsMatch(typeName) || ApiResponseTypeRegex.IsMatch(typeName);
-    }
-
-    private static void GenerateObsoleteAttribute(OpenApiOperation operation, StringBuilder code)
-    {
-        if (operation.IsDeprecated)
-        {
-            code.AppendLine($"{Separator}{Separator}[System.Obsolete]");
-        }
-    }
-
-    private static void GenerateForMultipartFormData(CSharpOperationModel operationModel, StringBuilder code)
-    {
-        if (operationModel.Consumes.Contains("multipart/form-data"))
-        {
-            code.AppendLine($"{Separator}{Separator}[Multipart]");
-        }
-    }
-
-    private void GenerateHeaders(
-        OpenApiOperation operation,
-        CSharpOperationModel operationModel,
-        StringBuilder code)
-    {
-        var headers = new List<string>();
-
-        if (settings.AddAcceptHeaders && document.SchemaType is >= NJsonSchema.SchemaType.OpenApi3)
-        {
-            var uniqueContentTypes = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-            foreach (var response in operation.Responses.Values)
-            {
-                if (response.Content == null)
-                    continue;
-
-                foreach (var contentType in response.Content.Keys)
-                {
-                    uniqueContentTypes.Add(contentType);
-                }
-            }
-
-            if (uniqueContentTypes.Any())
-            {
-                headers.Add($"\"Accept: {string.Join(", ", uniqueContentTypes)}\"");
-            }
-        }
-
-        if (settings.AddContentTypeHeaders && document.SchemaType is >= NJsonSchema.SchemaType.OpenApi3)
-        {
-            var uniqueContentTypes = operation.RequestBody?.Content.Keys ?? Array.Empty<string>();
-            var contentType =
-                uniqueContentTypes.FirstOrDefault(c => c.Equals("application/json", StringComparison.OrdinalIgnoreCase)) ??
-                uniqueContentTypes.FirstOrDefault();
-
-            if (!string.IsNullOrWhiteSpace(contentType) && !operationModel.Consumes.Contains("multipart/form-data"))
-            {
-                headers.Add($"\"Content-Type: {contentType}\"");
-            }
-        }
-
-        if (settings.AuthenticationHeaderStyle == AuthenticationHeaderStyle.Method)
-        {
-            foreach (var securitySchemeName in operationModel.Security.SelectMany(x => x.Keys))
-            {
-                if ((settings.SecurityScheme != null && securitySchemeName != settings.SecurityScheme) ||
-                    !document.SecurityDefinitions.TryGetValue(securitySchemeName, out var securityScheme))
-                {
-                    continue;
-                }
-
-                if (securityScheme is { Type: OpenApiSecuritySchemeType.Http, Scheme: var scheme }
-                    && string.Equals(scheme, "bearer", StringComparison.OrdinalIgnoreCase))
-                {
-                    headers.Add("\"Authorization: Bearer\"");
-                }
-            }
-        }
-
-        if (headers.Any())
-        {
-            code.AppendLine($"{Separator}{Separator}[Headers({string.Join(", ", headers)})]");
-        }
     }
 
     private string GenerateInterfaceDeclaration(string interfaceName, bool isSingleInterface)
